+++ run_tests.py
@@ -1,12 +1,14 @@
 import unittest
 
-from tests.auth_test import AuthTest
-from tests.profile_test import ProfileTest
-from tests.signup_test import SignupTest
-from tests.change_password_test import ChangePasswordTest
+from tests.profile.auth_test import AuthTest
+from tests.profile.profile_test import ProfileTest
+from tests.profile.signup_test import SignupTest
+from tests.profile.change_password_test import ChangePasswordTest
 
-from tests.main.main_test import MainTest
+from tests.main.messages_test import MessagesTest
 from tests.main.editor_test import EditorTest
+from tests.main.dialogues_test import DialoguesTest
+from tests.main.folders_test import FoldersTest
 
 if __name__ == '__main__':
     suite = unittest.TestSuite((
@@ -15,7 +17,9 @@ if __name__ == '__main__':
         unittest.makeSuite(SignupTest),
         unittest.makeSuite(ChangePasswordTest),
 
-        unittest.makeSuite(MainTest),
-        unittest.makeSuite(EditorTest)
+        unittest.makeSuite(MessagesTest),
+        unittest.makeSuite(EditorTest),
+        unittest.makeSuite(DialoguesTest),
+        unittest.makeSuite(FoldersTest)
     ))
     result = unittest.TextTestRunner().run(suite)
