+++ settings.template.py
@@ -1,5 +1,7 @@
 HUB_URL = 'http://127.0.0.1:4444/wd/hub'
 
+BROWSER = 'CHROME'  # CHROME or FIREFOX
+
 USERNAME = ''
 PASSWORD = ''
